+++ changelog.md
@@ -1,1 +1,1 @@
-* Fixed Colored Tiles don't cull [#374](https://github.com/Brandcraf06/Blockus/issues/374)
+* Fixed missing asphalt stairs/slab, netherite stairs/slab, cut soul sandstone slab, smooth stone stairs & ice brick wall in stairs, slabs & wall tags.